# stroom-auth-service-api/src/main/java/stroom/auth/service/api/model/SearchResponse.java
/*
 * Stroom Auth API
 * Various APIs for interacting with authentication, users, and tokens.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package stroom.auth.service.api.model;

import java.util.Objects;
import com.google.gson.annotations.SerializedName;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import stroom.auth.service.api.model.Token;

/**
 * A response for a search of tokens.
 */
@ApiModel(description = "A response for a search of tokens.")
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2017-11-01T10:45:36.027Z")
public class SearchResponse {
  @SerializedName("tokens")
  private List<Token> tokens = new ArrayList<Token>();

  @SerializedName("totalPages")
  private Integer totalPages = null;

  public SearchResponse tokens(List<Token> tokens) {
    this.tokens = tokens;
    return this;
  }

  public SearchResponse addTokensItem(Token tokensItem) {
    this.tokens.add(tokensItem);
    return this;
  }

   /**
   * A page of results.
   * @return tokens
  **/
  @ApiModelProperty(example = "null", required = true, value = "A page of results.")
  public List<Token> getTokens() {
    return tokens;
  }

  public void setTokens(List<Token> tokens) {
    this.tokens = tokens;
  }

  public SearchResponse totalPages(Integer totalPages) {
    this.totalPages = totalPages;
    return this;
  }

   /**
   * The total number of pages.
   * @return totalPages
  **/
  @ApiModelProperty(example = "null", required = true, value = "The total number of pages.")
  public Integer getTotalPages() {
    return totalPages;
  }

  public void setTotalPages(Integer totalPages) {
    this.totalPages = totalPages;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    SearchResponse searchResponse = (SearchResponse) o;
    return Objects.equals(this.tokens, searchResponse.tokens) &&
        Objects.equals(this.totalPages, searchResponse.totalPages);
  }

  @Override
  public int hashCode() {
    return Objects.hash(tokens, totalPages);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class SearchResponse {\n");
    
    sb.append("    tokens: ").append(toIndentedString(tokens)).append("\n");
    sb.append("    totalPages: ").append(toIndentedString(totalPages)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
  
}