# swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Various APIs for interacting with authentication, users, and tokens."
  version: "v1"
  title: "Stroom Auth API"
  contact:
    name: "GCHQ"
    url: "https://github.com/gchq/stroom-auth"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /token/v1:
    post:
      tags:
      - "ApiKey"
      summary: "Create a new token."
      description: ""
      operationId: "create"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CreateTokenRequest"
        required: false
        schema:
          $ref: "#/definitions/CreateTokenRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "integer"
            format: "int32"
    delete:
      tags:
      - "ApiKey"
      summary: "Delete all tokens."
      description: ""
      operationId: "deleteAll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /token/v1/byToken/{token}:
    get:
      tags:
      - "ApiKey"
      summary: "Read a token by the token string itself."
      description: ""
      operationId: "read"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Token"
    delete:
      tags:
      - "ApiKey"
      summary: "Delete a token by the token string itself."
      description: ""
      operationId: "delete"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /token/v1/search:
    post:
      tags:
      - "ApiKey"
      summary: "Submit a search request for tokens"
      description: ""
      operationId: "search"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "SearchRequest"
        required: false
        schema:
          $ref: "#/definitions/SearchRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SearchResponse"
  /token/v1/{id}:
    get:
      tags:
      - "ApiKey"
      summary: "Read a token by ID."
      description: ""
      operationId: "read"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Token"
    delete:
      tags:
      - "ApiKey"
      summary: "Delete a token by ID."
      description: ""
      operationId: "delete"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /token/v1/{id}/state:
    get:
      tags:
      - "ApiKey"
      summary: "Enable or disable the state of a token."
      description: ""
      operationId: "toggleEnabled"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - name: "enabled"
        in: "query"
        required: true
        type: "boolean"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
...
definitions:
  CreateTokenRequest:
    type: "object"
    required:
    - "tokenType"
    - "userEmail"
    properties:
      userEmail:
        type: "string"
        description: "The email of the user whom the token is for."
      tokenType:
        type: "string"
        description: "The type of token to create: e.g. user, api, or email_reset."
        pattern: "^user$|^api$|^email_reset$"
      comments:
        type: "string"
        description: "Comments about the token."
      enabled:
        type: "boolean"
        description: "Whether or not the new token should be enabled."
      parsedTokenType:
        type: "string"
        enum:
        - "USER"
        - "API"
        - "EMAIL_RESET"
    description: "A request to create a token."
...
  SearchRequest:
    type: "object"
    required:
    - "limit"
    - "page"
    properties:
      page:
        type: "integer"
        format: "int32"
        description: "The page of search results to retrieve."
      limit:
        type: "integer"
        format: "int32"
        description: "The number of tokens in a page of search results."
      orderBy:
        type: "string"
        description: "The property by which to order the results."
        pattern: "^enabled$|^user_email$|^issued_by_user$|^token$|^token_type$|^updated_by_user$|^expires_on$|^issued_on$|^updated_on$"
      orderDirection:
        type: "string"
        description: "The direction in which to order the results."
        pattern: "^asc$|^desc$"
      filters:
        type: "object"
        description: "How to filter the results. This is done by property, e.g. user_email,\
          \ 'someone@someplace.com'."
        additionalProperties:
          type: "string"
    description: "A request for a search over tokens."
  SearchResponse:
    type: "object"
    required:
    - "tokens"
    - "totalPages"
    properties:
      tokens:
        type: "array"
        description: "A page of results."
        items:
          $ref: "#/definitions/Token"
      totalPages:
        type: "integer"
        format: "int32"
        description: "The total number of pages."
    description: "A response for a search of tokens."
  Token:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      user_email:
        type: "string"
      token_type:
        type: "string"
      token:
        type: "string"
      expires_on:
        type: "string"
      issued_on:
        type: "string"
      issued_by_user:
        type: "string"
      enabled:
        type: "boolean"
      updated_on:
        type: "string"
      updated_by_user:
        type: "string"
...